Treat gid's identically to uid's for purposes of numeric formatting.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@816388 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/arch/unix/mod_unixd.c
@@ -115,8 +115,8 @@ static int set_group_privs(void)
          */
         if (setgid(ap_unixd_config.group_id) == -1) {
             ap_log_error(APLOG_MARK, APLOG_ALERT, errno, NULL,
-                        "setgid: unable to set group id to Group %u",
-                        (unsigned)ap_unixd_config.group_id);
+                        "setgid: unable to set group id to Group %ld",
+                        (long)ap_unixd_config.group_id);
             return -1;
         }
 
@@ -125,7 +125,7 @@ static int set_group_privs(void)
         if (initgroups(name, ap_unixd_config.group_id) == -1) {
             ap_log_error(APLOG_MARK, APLOG_ALERT, errno, NULL,
                         "initgroups: unable to set groups for User %s "
-                        "and Group %u", name, (unsigned)ap_unixd_config.group_id);
+                        "and Group %ld", name, (long)ap_unixd_config.group_id);
             return -1;
         }
 #endif /* !defined(OS2) */
